requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
certifi
cffi
charset-normalizer
click==8.1.7
comm
contourpy==1.1.1
coverage==7.6.1
cycler==0.12.1
debugpy
decorator
deepspeed==0.14.4
dictor==0.1.12
docker-pycreds==0.4.0
dotmap==1.3.30
einops==0.8.0
elementpath==1.3.3
ephem==4.1.5
executing==2.0.1
filelock==3.0.12
filterpy==1.4.5
flash-attn==2.6.3
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h2
hjson==3.1.0
hpack==4.0.0
huggingface-hub==0.24.5
hydra-core==1.3.2
hyperframe
idna
imageio==2.35.1
imgaug==0.4.0
importlib_metadata
importlib_resources==6.4.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.12.3
jedi
Jinja2
joblib==1.4.2
jsonpickle==4.0.2
jupyter_client
jupyter_core
kiwisolver==1.4.5
laspy==2.5.4
lazy_loader==0.4
line_profiler==4.2.0
MarkupSafe
matplotlib==3.5.3
matplotlib-inline
mpmath
natsort==8.4.0
nest_asyncio
networkx
ninja==1.11.1.1
numpy
nvidia-ml-py==12.555.43
omegaconf==2.3.0
opencv-python==4.2.0.32
packaging==24.1
pandas==2.0.3
parso
pexpect
pickleshare
Pillow
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil
ptyprocess
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-trees==0.8.3
pycparser
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
pygame==2.6.0
Pygments
pyparsing==3.1.2
PySocks
python-dateutil
pytz==2024.1
PyWavelets==1.4.1
PyYAML
pyzmq
rdp==0.8
regex==2024.7.24
requests
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
Shapely==1.7.1
simple_watchdog_timer==0.1.1
six
smmap==5.0.1
stack-data==0.6.3
sympy
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado
tqdm==4.66.5
traitlets
transformers==4.43.4
transforms3d==0.4.2
triton==3.0.0
typing_extensions
tzdata==2024.1
ujson==5.10.0
urllib3
wandb==0.17.5
wcwidth
xmlschema==1.0.18
zipp==3.20.2
zstandard==0.23.0